requirements.txt
django
django-json-rpc
django-annoying
stormed
-e git+https://github.com/facebook/tornado.git#egg=tornado
-e git+https://github.com/mrjoes/sockjs-tornado.git#egg=sockjs_tornado
